Issue137-server-cgi実行

+++ src/server/NetworkIOHandler.hpp
@@ -51,7 +51,7 @@ class NetworkIOHandler
 		int	setupSocket( const std::string address, const unsigned int port );
 		int receiveRequest( ConnectionManager& connManager, const int cli_sock );
 		int sendResponse( ConnectionManager& connManager, const int cli_sock );
-		ssize_t receiveCgiResponse( ConnectionManager& connManager, const int sock );
+		int receiveCgiResponse( ConnectionManager& connManager, const int sock );
 		ssize_t sendRequestBody( ConnectionManager& connManager, const int sock );
 		int acceptConnection( ConnectionManager& connManager, const int listen_fd );
 		void closeConnection( ConnectionManager& connManager, const int cli_sock );
